+++ client/index.ts
@@ -3,3 +3,5 @@ import validateConfig from '../packages/validate-config'
 export { ReferenceDataStore } from './src/reference-data'
 export * as definitionService from './src/definition-manipulation'
 export { validateConfig }
+export { useLogin } from './src/hooks/useLogin'
+export { useLogout } from './src/hooks/useLogout'