+++ client/src/hooks/useLogin.ts
@@ -1,0 +1,8 @@
+import { authService } from '@oneblink/apps'
+import { useCallback } from 'react'
+
+export function useLogin(cognitoClientID: string) {
+  return useCallback(async () => {
+    await authService.loginHostedUI(cognitoClientID)
+  }, [])
+}