+++ client/src/hooks/useLogout.ts
@@ -1,0 +1,9 @@
+import { authService } from '@oneblink/apps'
+import { useCallback } from 'react'
+
+export function useLogout() {
+  return useCallback(async () => {
+    authService.logoutHostedUI()
+    await authService.logout()
+  }, [])
+}