+++ CHANGELOG.md
@@ -6,4 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
+### Added
+
 - Validate Config function
+- useLogin/useLogout hooks